% Please edit documentation in R/chspider.batch.annotation.child.R
\name{chspider.batch.annotation.child}
\alias{chspider.batch.annotation.child}
\title{Human Metabolome Database', 'ChEMBL', 'ChEBI', 'NIAID','Pesticide Common Names','SMPDB Small Molecule Pathway Database',
MeSH','LipidMAPS','ChemBank','BioCyc'),token=tokenstr)}
\usage{
chspider.batch.annotation.child(mz.val, max.mz.diff, adductname, 
    datasources = c("KEGG"), tokenstr, maxhits = 30, syssleep = 10, 
    adduct_table)
}
\description{
Human Metabolome Database', 'ChEMBL', 'ChEBI', 'NIAID','Pesticide Common Names','SMPDB Small Molecule Pathway Database',
MeSH','LipidMAPS','ChemBank','BioCyc'),token=tokenstr)
}
