% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.annotationvold.R
\name{KEGG.annotationvold}
\alias{KEGG.annotationvold}
\title{KEGG.Gene.ID',}
\usage{
KEGG.annotationvold(dataA, max.mz.diff = 10, queryadductlist = c("M+H"),
  xMSannotator.outloc, numnodes = 1, syssleep = 1, adduct_table)
}
\description{
KEGG.Gene.ID',
}
