% Please edit documentation in R/Metlin.annotation.R
\name{Metlin.annotation}
\alias{Metlin.annotation}
\title{Human Metabolome Database', 'ChEMBL', 'ChEBI', 'NIAID','Pesticide Common Names','SMPDB Small Molecule Pathway Database',
MeSH','LipidMAPS','ChemBank','BioCyc')}
\usage{
Metlin.annotation(dataA, max.mz.diff = 10, queryadductlist = c("M+H"), 
    xMSannotator.outloc, tokenstr = NA)
}
\description{
Human Metabolome Database', 'ChEMBL', 'ChEBI', 'NIAID','Pesticide Common Names','SMPDB Small Molecule Pathway Database',
MeSH','LipidMAPS','ChemBank','BioCyc')
}
